Mobile Operating System Model in C Plus Plus
Mobile Operating System Model in C++

This project has been developed in C++ , graphics are purely from the
BGI file of Turbo C++.The compiler used is Turbo C++.

How to use this ?

Copy the whole script and paste it in a notepad or in any of your
favorite IDE . Save it as mobOs.cpp (extension must be cpp) . Run it
from Turbo C++ Compiler (Recommended ) . In other compilers the graphics
may not be compaitable therefore Turbo C++ is recommended.

Use default username = hunter and password = hunter

Navigate Through the Operating System and play with different tools.

Get ideas and spread the knowledge with your friends , because sharing
is caring!

Thanks for your time!